Backend/requirements.txt
# requirements.txt
flask
firebase_admin
google-cloud-secret-manager==2.0.0
python-dotenv
flask-cors
ariadne
gunicorn
Authlib==0.15.3